+++ src/main/java/com/kelseyde/calvin/search/picker/MovePicker.java
@@ -1,4 +1,4 @@
-package com.kelseyde.calvin.generation.picker;
+package com.kelseyde.calvin.search.picker;
 
 import com.kelseyde.calvin.board.Board;
 import com.kelseyde.calvin.board.Move;
@@ -11,52 +11,59 @@ import lombok.experimental.FieldDefaults;
 import java.util.List;
 
 @FieldDefaults(level = AccessLevel.PRIVATE)
-public class QuiescentMovePicker implements MovePicking {
+public class MovePicker implements MovePicking {
 
     public enum Stage {
-        PREVIOUS_BEST_MOVE,
+        BEST_MOVE,
         NOISY,
+        QUIET,
         END
     }
 
     final MoveGeneration moveGenerator;
     final MoveOrdering moveOrderer;
 
-    final Board board;
-    MoveFilter filter;
+    Board board;
+    int ply;
+
     Stage stage;
 
     List<Move> moves;
-    Move previousBestMove;
+    Move bestMove;
     int moveIndex;
     int[] scores;
 
-    public QuiescentMovePicker(MoveGeneration moveGenerator, MoveOrdering moveOrderer, Board board) {
+    public MovePicker(MoveGeneration moveGenerator, MoveOrdering moveOrderer, Board board, int ply) {
         this.moveGenerator = moveGenerator;
         this.moveOrderer = moveOrderer;
         this.board = board;
-        this.stage = Stage.PREVIOUS_BEST_MOVE;
+        this.ply = ply;
+        this.stage = Stage.BEST_MOVE;
     }
 
     @Override
     public Move pickNextMove() {
 
         Move nextMove = null;
-
         while (nextMove == null) {
             nextMove = switch (stage) {
-                case PREVIOUS_BEST_MOVE -> pickPreviousBestMove();
-                case NOISY -> pickMove();
+                case BEST_MOVE -> pickPreviousBestMove();
+                case NOISY -> pickMove(MoveFilter.NOISY, Stage.QUIET);
+                case QUIET -> pickMove(MoveFilter.QUIET, Stage.END);
                 case END -> null;
             };
             if (stage == Stage.END) break;
         }
-
         return nextMove;
 
     }
 
-    private Move pickMove() {
+    private Move pickPreviousBestMove() {
+        stage = Stage.NOISY;
+        return bestMove;
+    }
+
+    private Move pickMove(MoveFilter filter, Stage nextStage) {
 
         if (moves == null) {
             moves = moveGenerator.generateMoves(board, filter);
@@ -65,29 +72,24 @@ public class QuiescentMovePicker implements MovePicking {
         }
         if (moveIndex >= moves.size()) {
             moves = null;
-            stage = Stage.END;
+            stage = nextStage;
             return null;
         }
         sortMoves();
         Move move = moves.get(moveIndex);
         moveIndex++;
-        if (move.equals(previousBestMove)) {
+        if (move.equals(bestMove)) {
             // Skip to the next move
-            return pickMove();
+            return pickMove(filter, nextStage);
         }
         return move;
 
     }
 
-    private Move pickPreviousBestMove() {
-        stage = Stage.NOISY;
-        return previousBestMove;
-    }
-
     public void scoreMoves() {
         scores = new int[moves.size()];
         for (int i = 0; i < moves.size(); i++) {
-            scores[i] = moveOrderer.mvvLva(board, moves.get(i), previousBestMove);
+            scores[i] = moveOrderer.scoreMove(board, moves.get(i), bestMove, ply);
         }
     }
 
@@ -110,13 +112,13 @@ public class QuiescentMovePicker implements MovePicking {
         return moves;
     }
 
-
-    public void setPreviousBestMove(Move previousBestMove) {
-        this.previousBestMove = previousBestMove;
+    public void setMoves(List<Move> moves) {
+        this.moves = moves;
+        scoreMoves();
     }
 
-    public void setFilter(MoveFilter filter) {
-        this.filter = filter;
+    public void setBestMove(Move bestMove) {
+        this.bestMove = bestMove;
     }
 
 }
