+++ src/main/java/com/kelseyde/calvin/tuning/perft/PerftService.java
@@ -3,8 +3,6 @@ package com.kelseyde.calvin.tuning.perft;
 import com.kelseyde.calvin.board.Board;
 import com.kelseyde.calvin.board.Move;
 import com.kelseyde.calvin.generation.MoveGenerator;
-import com.kelseyde.calvin.generation.picker.MovePicker;
-import com.kelseyde.calvin.search.moveordering.MoveOrderer;
 import com.kelseyde.calvin.utils.notation.Notation;
 
 import java.util.List;
@@ -13,17 +11,14 @@ import java.util.Set;
 public class PerftService {
 
     private final MoveGenerator moveGenerator = new MoveGenerator();
-    private final MoveOrderer moveOrderer = new MoveOrderer();
 
     public long perft(Board board, int depth) {
-        MovePicker movePicker = new MovePicker(moveGenerator, moveOrderer, board, depth);
+        List<Move> moves = moveGenerator.generateMoves(board);
         if (depth == 1) {
-            return moveGenerator.generateMoves(board).size();
+            return moves.size();
         }
         long totalMoveCount = 0;
-        while (true) {
-            Move move = movePicker.pickNextMove();
-            if (move == null) break;
+        for (Move move : moves) {
             board.makeMove(move);
             totalMoveCount += perft(board, depth - 1);
             board.unmakeMove();
