+++ src/test/java/com/kelseyde/calvin/generation/MovePickerTest.java
@@ -2,7 +2,7 @@ package com.kelseyde.calvin.generation;
 
 import com.kelseyde.calvin.board.Board;
 import com.kelseyde.calvin.board.Move;
-import com.kelseyde.calvin.generation.picker.MovePicker;
+import com.kelseyde.calvin.search.picker.MovePicker;
 import com.kelseyde.calvin.search.moveordering.MoveOrderer;
 import com.kelseyde.calvin.utils.TestUtils;
 import com.kelseyde.calvin.utils.notation.FEN;
@@ -30,40 +30,11 @@ public class MovePickerTest {
 
         movePicker = new MovePicker(moveGenerator, moveOrderer, board, 1);
 
-        movePicker.setPreviousBestMove(previousBestMove);
+        movePicker.setBestMove(previousBestMove);
 
         Move move = movePicker.pickNextMove();
 
         Assertions.assertNull(movePicker.getMoves());
-
-    }
-
-    @Test
-    @Disabled
-    public void testAll() throws IOException {
-
-        List<Board> positions = TestUtils.loadFens().stream()
-                .map(FEN::toBoard)
-                .toList()
-                .subList(0, 10000);
-
-        for (Board board : positions) {
-            List<Move> moves = moveGenerator.generateMoves(board);
-
-            List<Move> orderedMoves = moveOrderer.orderMoves(board, moves, null, 0);
-
-            movePicker = new MovePicker(moveGenerator, moveOrderer, board, 0);
-            List<Move> pickedMoves = new ArrayList<>();
-            while (true) {
-                Move move = movePicker.pickNextMove();
-                if (move == null) break;
-                pickedMoves.add(move);
-            }
-
-            Assertions.assertEquals(orderedMoves.size(), pickedMoves.size());
-            Assertions.assertEquals(orderedMoves, pickedMoves);
-        }
-
 
     }
